Ryan's feedback moving classes out of server package
