+++ extensions/barrage/src/main/java/io/deephaven/extensions/barrage/util/BarrageUtil.java
@@ -9,10 +9,22 @@ import com.google.protobuf.ByteStringAccess;
 import com.google.rpc.Code;
 import io.deephaven.UncheckedDeephavenException;
 import io.deephaven.base.ClassUtil;
+import io.deephaven.engine.rowset.RowSequence;
+import io.deephaven.engine.rowset.RowSet;
+import io.deephaven.engine.rowset.RowSetFactory;
+import io.deephaven.engine.rowset.WritableRowSet;
 import io.deephaven.engine.table.ColumnDefinition;
 import io.deephaven.engine.table.Table;
 import io.deephaven.engine.table.TableDefinition;
 import io.deephaven.engine.table.MatchPair;
+import io.deephaven.engine.table.impl.BaseTable;
+import io.deephaven.engine.table.impl.remote.ConstructSnapshot;
+import io.deephaven.engine.table.impl.util.BarrageMessage;
+import io.deephaven.engine.updategraph.UpdateGraphProcessor;
+import io.deephaven.extensions.barrage.BarrageMessageProducer;
+import io.deephaven.extensions.barrage.BarragePerformanceLog;
+import io.deephaven.extensions.barrage.BarrageSnapshotOptions;
+import io.deephaven.extensions.barrage.BarrageStreamGenerator;
 import io.deephaven.extensions.barrage.chunk.vector.VectorExpansionKernel;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
@@ -28,6 +40,7 @@ import io.deephaven.chunk.ChunkType;
 import io.deephaven.proto.backplane.grpc.ExportedTableCreationResponse;
 import io.deephaven.util.type.TypeUtils;
 import io.deephaven.vector.Vector;
+import io.grpc.stub.StreamObserver;
 import org.apache.arrow.flatbuf.KeyValue;
 import org.apache.arrow.util.Collections2;
 import org.apache.arrow.vector.types.TimeUnit;
@@ -44,6 +57,7 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.BitSet;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -57,7 +71,14 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.IntFunction;
 
+import static io.deephaven.extensions.barrage.BarrageMessageProducer.MAX_SNAPSHOT_CELL_COUNT;
+import static io.deephaven.extensions.barrage.BarrageMessageProducer.MIN_SNAPSHOT_CELL_COUNT;
+import static io.deephaven.extensions.barrage.BarrageMessageProducer.TARGET_SNAPSHOT_PERCENTAGE;
+
 public class BarrageUtil {
+    public static final BarrageSnapshotOptions DEFAULT_SNAPSHOT_DESER_OPTIONS =
+            BarrageSnapshotOptions.builder().build();
+
     public static final long FLATBUFFER_MAGIC = 0x6E687064;
 
     private static final Logger log = LoggerFactory.getLogger(BarrageUtil.class);
@@ -589,5 +610,145 @@ public class BarrageUtil {
                 "", componentType, componentType.getComponentType(), Collections.emptyMap()));
 
         return new Field(name, fieldType, children);
+    }
+
+    public static void createAndSendStaticSnapshot(
+            BarrageMessageProducer.StreamGenerator.Factory<BarrageStreamGenerator.View> streamGeneratorFactory,
+            BaseTable table,
+            BitSet columns,
+            RowSet viewport,
+            boolean reverseViewport,
+            BarrageSnapshotOptions snapshotRequestOptions,
+            StreamObserver<BarrageStreamGenerator.View> listener,
+            BarragePerformanceLog.SnapshotMetricsHelper metrics) {
+        // start with small value and grow
+        long snapshotTargetCellCount = MIN_SNAPSHOT_CELL_COUNT;
+        double snapshotNanosPerCell = 0.0;
+
+        final long columnCount =
+                Math.max(1, columns != null ? columns.cardinality() : table.getDefinition().getColumns().size());
+
+        try (final WritableRowSet snapshotViewport = RowSetFactory.empty();
+                final WritableRowSet targetViewport = RowSetFactory.empty()) {
+            // compute the target viewport
+            if (viewport == null) {
+                targetViewport.insertRange(0, table.size() - 1);
+            } else if (!reverseViewport) {
+                targetViewport.insert(viewport);
+            } else {
+                // compute the forward version of the reverse viewport
+                try (final RowSet rowKeys = table.getRowSet().subSetForReversePositions(viewport);
+                        final RowSet inverted = table.getRowSet().invert(rowKeys)) {
+                    targetViewport.insert(inverted);
+                }
+            }
+
+            try (final RowSequence.Iterator rsIt = targetViewport.getRowSequenceIterator()) {
+                while (rsIt.hasMore()) {
+                    // compute the next range to snapshot
+                    final long cellCount =
+                            Math.max(MIN_SNAPSHOT_CELL_COUNT,
+                                    Math.min(snapshotTargetCellCount, MAX_SNAPSHOT_CELL_COUNT));
+
+                    final RowSequence snapshotPartialViewport = rsIt.getNextRowSequenceWithLength(cellCount);
+                    // add these ranges to the running total
+                    snapshotPartialViewport.forEachRowKeyRange((start, end) -> {
+                        snapshotViewport.insertRange(start, end);
+                        return true;
+                    });
+
+                    // grab the snapshot and measure elapsed time for next projections
+                    long start = System.nanoTime();
+                    final BarrageMessage msg =
+                            ConstructSnapshot.constructBackplaneSnapshotInPositionSpace(log, table,
+                                    columns, snapshotPartialViewport, null);
+                    msg.modColumnData = BarrageMessage.ZERO_MOD_COLUMNS; // no mod column data for DoGet
+                    long elapsed = System.nanoTime() - start;
+                    // accumulate snapshot time in the metrics
+                    metrics.snapshotNanos += elapsed;
+
+                    // send out the data. Note that although a `BarrageUpdateMetaData` object will
+                    // be provided with each unique snapshot, vanilla Flight clients will ignore
+                    // these and see only an incoming stream of batches
+                    try (final BarrageMessageProducer.StreamGenerator<BarrageStreamGenerator.View> bsg =
+                            streamGeneratorFactory.newGenerator(msg, metrics)) {
+                        if (rsIt.hasMore()) {
+                            listener.onNext(bsg.getSnapshotView(snapshotRequestOptions,
+                                    snapshotViewport, false,
+                                    msg.rowsIncluded, columns));
+                        } else {
+                            listener.onNext(bsg.getSnapshotView(snapshotRequestOptions,
+                                    viewport, reverseViewport,
+                                    msg.rowsIncluded, columns));
+                        }
+                    }
+
+                    if (msg.rowsIncluded.size() > 0) {
+                        // very simplistic logic to take the last snapshot and extrapolate max
+                        // number of rows that will not exceed the target UGP processing time
+                        // percentage
+                        long targetNanos = (long) (TARGET_SNAPSHOT_PERCENTAGE
+                                * UpdateGraphProcessor.DEFAULT.getTargetCycleDurationMillis()
+                                * 1000000);
+
+                        long nanosPerCell = elapsed / (msg.rowsIncluded.size() * columnCount);
+
+                        // apply an exponential moving average to filter the data
+                        if (snapshotNanosPerCell == 0) {
+                            snapshotNanosPerCell = nanosPerCell; // initialize to first value
+                        } else {
+                            // EMA smoothing factor is 0.1 (N = 10)
+                            snapshotNanosPerCell =
+                                    (snapshotNanosPerCell * 0.9) + (nanosPerCell * 0.1);
+                        }
+
+                        snapshotTargetCellCount =
+                                (long) (targetNanos / Math.max(1, snapshotNanosPerCell));
+                    }
+                }
+            }
+        }
+    }
+
+    public static void createAndSendSnapshot(
+            BarrageMessageProducer.StreamGenerator.Factory<BarrageStreamGenerator.View> streamGeneratorFactory,
+            BaseTable table,
+            BitSet columns, RowSet viewport, boolean reverseViewport,
+            BarrageSnapshotOptions snapshotRequestOptions,
+            StreamObserver<BarrageStreamGenerator.View> listener,
+            BarragePerformanceLog.SnapshotMetricsHelper metrics) {
+
+        // if the table is static and a full snapshot is requested, we can make and send multiple
+        // snapshots to save memory and operate more efficiently
+        if (!table.isRefreshing()) {
+            createAndSendStaticSnapshot(streamGeneratorFactory, table, columns, viewport, reverseViewport,
+                    snapshotRequestOptions, listener, metrics);
+            return;
+        }
+
+        // otherwise snapshot the entire request and send to the client
+        final BarrageMessage msg;
+
+        final long snapshotStartTm = System.nanoTime();
+        if (reverseViewport) {
+            msg = ConstructSnapshot.constructBackplaneSnapshotInPositionSpace(log, table,
+                    columns, null, viewport);
+        } else {
+            msg = ConstructSnapshot.constructBackplaneSnapshotInPositionSpace(log, table,
+                    columns, viewport, null);
+        }
+        metrics.snapshotNanos = System.nanoTime() - snapshotStartTm;
+
+        msg.modColumnData = BarrageMessage.ZERO_MOD_COLUMNS; // no mod column data
+
+        // translate the viewport to keyspace and make the call
+        try (final BarrageMessageProducer.StreamGenerator<BarrageStreamGenerator.View> bsg =
+                streamGeneratorFactory.newGenerator(msg, metrics);
+                final RowSet keySpaceViewport = viewport != null
+                        ? msg.rowsAdded.subSetForPositions(viewport, reverseViewport)
+                        : null) {
+            listener.onNext(bsg.getSnapshotView(
+                    snapshotRequestOptions, viewport, reverseViewport, keySpaceViewport, columns));
+        }
     }
 }
