+++ server/src/test/java/io/deephaven/server/barrage/BarrageStreamTableTest.java
@@ -27,13 +27,15 @@ import io.deephaven.engine.updategraph.UpdateGraphProcessor;
 import io.deephaven.engine.updategraph.UpdateSourceCombiner;
 import io.deephaven.engine.util.TableDiff;
 import io.deephaven.engine.util.TableTools;
+import io.deephaven.extensions.barrage.BarrageMessageProducer;
+import io.deephaven.extensions.barrage.BarrageStreamGenerator;
 import io.deephaven.extensions.barrage.BarrageSubscriptionOptions;
 import io.deephaven.extensions.barrage.table.BarrageTable;
 import io.deephaven.extensions.barrage.util.BarrageStreamReader;
 import io.deephaven.extensions.barrage.util.BarrageUtil;
 import io.deephaven.proto.flight.util.SchemaHelper;
 import io.deephaven.server.arrow.ArrowModule;
-import io.deephaven.server.util.Scheduler;
+import io.deephaven.util.Scheduler;
 import io.deephaven.server.util.TestControlledScheduler;
 import io.deephaven.test.types.OutOfBandTest;
 import io.deephaven.util.annotations.ReferentialIntegrity;
@@ -98,7 +100,7 @@ public class BarrageStreamTableTest extends RefreshingTableTestCase {
         streamTable.setRefreshing(true);
         streamTable.setAttribute(Table.STREAM_TABLE_ATTRIBUTE, true);
 
-        barrageMessageProducer = streamTable.getResult(new BarrageMessageProducer.Operation<>(
+        barrageMessageProducer = streamTable.getResult(new BarrageMessageProducerOperation<>(
                 scheduler, daggerRoot.getStreamGeneratorFactory(), streamTable, UPDATE_INTERVAL, () -> {
                 }));
